pyproject.toml
[project]
name = "rldc-trading-ainalyzer-bot"
version = "0.7.0-beta"
description = "RLdC Trading AiNalyzer Bot v0.7 beta"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "RLdC", email = "devnull@example.com" }]
license = { text = "MIT" }
keywords = ["trading", "analysis", "bot", "crypto", "ai"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
]

dependencies = [
  "apscheduler>=3.10.4",
  "ccxt>=4.3.0",
  "fastapi>=0.111.0",
  "jinja2>=3.1.3",
  "numpy>=1.26.4",
  "openai>=1.40.0",
  "pandas>=2.2.2",
  "pyarrow>=16.1.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "python-telegram-bot>=21.3",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "uvicorn>=0.30.1",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.4.2",
  "mypy>=1.10.1",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "types-python-dateutil>=2.9.0.20240316",
]

[project.scripts]
rldc = "rldc.cli:app"

[tool.black]
line-length = 100

target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
exclude = [".venv", "build", "dist"]

[tool.ruff.lint.isort]
known-first-party = ["rldc"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69.5", "wheel"]
build-backend = "setuptools.build_meta"
